apps/api/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP client (for OpenRouter)
httpx>=0.26.0
openai>=1.10.0
python-multipart>=0.0.9

# Environment
python-dotenv>=1.0.0

# Logging
structlog>=24.1.0

# Storage + DB
sqlalchemy>=2.0.0
boto3>=1.34.0

# STT
faster-whisper>=1.0.0
numpy>=1.26.0

# File parsing
pypdf>=4.0.0
python-docx>=1.1.0

# Dev dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
ruff>=0.2.0
mypy>=1.8.0
pyinstaller>=6.0.0
